Move grdfilter to src
