+++ pygmt/src/__init__.py
@@ -3,6 +3,7 @@ Source code for PyGMT modules.
 """
 # pylint: disable=import-outside-toplevel
 from pygmt.src.blockmedian import blockmedian
+from pygmt.src.grdfilter import grdfilter
 from pygmt.src.grdtrack import grdtrack
 from pygmt.src.makecpt import makecpt
 from pygmt.src.meca import meca
